+++ CHANGELOG.md
@@ -61,7 +61,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 
 ### New
 
-- TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
+- **Forgejo Scaler**: Provide support for actions and pending jobs on forgejo ([#6488](https://github.com/kedacore/keda/issues/6488))
 
 #### Experimental